Resolve @highlight-run/client without relative paths (#4851)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

_This is part of a
[series](https://github.com/highlight/highlight/pull/4813)
[of](https://github.com/highlight/highlight/pull/4848)
[PRs](https://github.com/highlight/highlight/pull/4849) being spun off
from [my WIP branch](https://github.com/lewisl9029/highlight/pull/2) to
get the Highlight web app ready for [Reflame](https://reflame.app/).
Hopefully this makes things a bit easier to review, test, and merge. 🙂_

The sdk/firstload package currently imports from `@highlight-run/client`
using relative paths that point outside the package root. I read your
[blog post on
this](https://www.highlight.io/blog/publishing-private-pnpm-monorepo)
and was able to get a good understanding of the motivations, but it was
breaking the app in Reflame (where individual libraries are much more
tightly isolated at the boundary, and relative imports outside the
package root would basically just point into the aether) so I had to
find a workaround. 😅

I think I found a reasonable one here, where we use the [`modulePaths`
option](https://github.com/rollup/plugins/tree/master/packages/node-resolve#modulepaths)
in `@rollup/plugin-node-resolve` to make `@highlight-run/client`
resolvable as an internal module, so it gets included in the bundle
without the need to specify it as a dependency in package.json.

The resulting builds ended up having identical hashes as before the
change, so this did manage to preserve the same output as before. Wdyt?

## How did you test this change?

I added `entryFileNames: '[name]-[hash].js'` to output options and ran
the build in this branch and main. Both ended up with identical hashes,
which means the build output was not affected at all by this change. 😄


<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

Since the build output is identical I think we're good, but it could be
a good idea to cut a new release just in case to confirm this doesn't
cause any unforeseen issues in the wild.