Issue #4641: React ErrorBoundary styles should be loaded automatically (#5031)
## Summary

Addresses issue #4641 

The `<ErrorBoundary>` component exported from `@highlight-run/react` did
not have self-contained styles. The user had to manually import
`@highlight-run/react/dist/index.css`.

## How did you test this change?

We've added an `<ErrorBoundaryButton />` to the `nextjs` test project,
and wired it up to the new `<ErrorBoundary>` component. Run `yarn
e2e:nextjs` from the project root to spin up the server on
`localhost:3005`. Click the "throw an error" button to trigger.


![image](https://user-images.githubusercontent.com/878947/233398725-314976ee-2e51-43cb-aef7-79ab782450c9.png)


![image](https://user-images.githubusercontent.com/878947/233400476-a13c4de1-6e6d-47c3-af93-b5dc9d22575b.png)

## Future considerations

Issue #4288 is a redesign of the `<ErrorBoundary showDialog>` dialog.
I'll reconsider the font situation during that effort.

## Are there any deployment considerations?

n/a