add a GitHub integration for creating issues (#5025)
## Summary

Builds [a GitHub app](https://github.com/apps/highlight-io/) for
Highlight for creating issues from comments and errors.
The GitHub app is installed on a Highlight workspace since the app is
installed to a GitHub workspace.
Currently for creating issues, we show all repositories that the app has
access to (within the org that it is installed to)
in the dropdown of the creation flow. In the future, the integration
settings modal could be used to connect the
app to particular highlight projects (eg if we want to associate a
highlight project with a repo for tracking deploys).

We use an app installation workflow to get access to the 'organization'
into which
an app is installed. if we went with an oauth workflow, we would only
have access
to the repositories the installing user owns. see [github
docs](https://docs.github.com/en/apps/creating-github-apps/authenticating-with-a-github-app/authenticating-as-a-github-app-installation)
for more details on the auth flow.

We use the existing `model.IntegrationWorkspaceMapping` workflow for
storing the integration state of
the GitHub app with a workspace.

## How did you test this change?

https://www.loom.com/share/29fe17a8d78f42d8bf2ca3ac1aaab982

## Are there any deployment considerations?

No