remove logic to fetch error instance that is not live (#5223)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

#3576 attempted to fetch the first error instance with a non-live
session (`processed=true`). The reasoning behind this was the devtools
in for livemode aren't great (#4965). Additionally, #4269 tried to avoid
error instances tied to excluded sessions.

The problem with these approaches is that the error group's last
instance timestamp doesn't match one these selected. Here's a visual:

![236049033-4e4d768f-a4ad-441b-b453-f815aba8b9a7](https://user-images.githubusercontent.com/58678/236330902-0e860977-0622-4fbc-a355-7b2d94d928f8.png)

This PR removes the backend logic and relies more on frontend UX to
inform the user if there is an issue.




## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

If the session is excluded, we disable the "Show session" button and use
a callout:
![Screenshot 2023-05-04 at 3 04 10
PM](https://user-images.githubusercontent.com/58678/236331027-13673129-a778-42d1-8c51-f8ebd5ec32cd.png)

If the session is live, we provide a tooltip to warn them about dev
tools:
![Kapture 2023-05-04 at 10 11
41](https://user-images.githubusercontent.com/58678/236331238-7b4a16b1-02d3-46b0-8304-46a5ea455c12.gif)


## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

Will confirm with the customer that reported this that this fixes their
issue.