withHighlightConfig doc improvement (#5257)
## Summary

Addresses #4057

We've added some comments to the `withHighlightConfig` docs. 

In short, `withHighlightConfig` returns a promise, which is supported
for Next v12.1.0 and greater. This breaks old Next apps as well as Next
config generator functions that haven't learned to accept a promise as
their argument. `next-pwa` is the offender in this case.


![image](https://github.com/highlight/highlight/assets/878947/8e1f10f9-7170-4247-9842-0e7a44cc9e6b)

## How did you test this change?

Spun up the docs.

## Are there any deployment considerations?

Nope