Request Headers (#5438)
## Summary

We've had consistent bug reports of missing headers.

Here's an example:
https://discord.com/channels/1026884757667188757/1109049760909885480

Huge shoutout to Xvipes on Discord for reporting this bug with a perfect
reproduction and even tracking down the line that needed fixing.

This likely closes a bunch of issues in the repo.

## How did you test this change?

I followed Xvipe's reproduction, using
[sindresorhus/ky](https://github.com/sindresorhus/ky) to make the
`fetch` request. `ky` uses the native `Request` object in a way that we
weren't supporting.

## Are there any deployment considerations?

Nope! Just a patch version bump.