fix incorrect usage of lazy query (#5127)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

There's an interesting bug whereby we can't scroll when changing the
`query`. To reproduce:

* Visit the logs page and set the time frame to something small (Last 15
minutes is sufficient for local, you may need something smaller for
prod)
* You should have more than one page of logs
* Set the query to: `level:trace`
* Keep scrolling until you hit the end of all trace logs
* Clear the query
* Observe that you cannot scroll past the first page of logs

The reason this was happening was because I initially configured the
lazy query in #4535 to apply the variables at the instantiation of the
lazy query, not when I actually call the query (`getLogs()`). This is a
gotcha with apollo (see
[#5912](https://github.com/apollographql/apollo-client/issues/5912)).

The reason why we even use a lazy query in the first place is because
changing the window info is dependent on what we're doing:
* first time the page loads, set window info from the results of
invoking the lazy query
* variable (such as `query`, start time, end time) changes, set window
info from the results of invoking the lazy query
* scrolling to the next page (or previous page), set window info based
on the
[`fetchMore`](https://www.apollographql.com/docs/react/pagination/core-api/#the-fetchmore-function)
function

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Per the test above, we should be able to scroll past that first page of
logs after clearing the query.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A