make all public resolver opensearch updates be synchronous (#5524)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

We are noticing that sometimes opensearch can be out of sync with
postgres (#5511, #5454, #5301) due to the fact that the default behavior
is to perform the updates asynchronously.

Per this
[slack](https://highlightcorp.slack.com/archives/C02CJANPHQS/p1685660764377189)
conversation, we would like to make all opensearch updates synchronous.
The reasoning behind doing async updates predates having kafka to
throttle updates. Now that we have kafka, we would like to switch to
only doing sync updates to see if that fixes our opensearch data quality
problems.

Since the public resolver is where we'll experience the most amount of
updates, we want to make this change independently so we can monitor the
effect it has on our opensearch cluster (and roll back if it's
problematic).

## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

Confirmed opensearch updates that are sent via the public graph still
work as expected (e.g. consuming a new error).

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

Will monitor the [cluster
health](https://us-east-2.console.aws.amazon.com/aos/home?region=us-east-2#opensearch/domains/highlight-search-prod-v2?tabId=clusterHealth)
after this roll out.